feat(linux): remove filesystem.h/cpp and move OpenFile() function to mc_kmx_file
