+++ linux/mcompile/keymap/mc_kmxfile.h
@@ -4,7 +4,6 @@
 
 #include "km_types.h"
 #include <kmx_file.h>
-#include "filesystem.h"
 #include "mcompile.h"
 
 #ifndef _KMXFILE_H
@@ -74,6 +73,11 @@ KMX_BOOL KMX_LoadKeyboard(char16_t* fileName, LPKMX_KEYBOARD *lpKeyboard);
 KMX_BOOL KMX_SaveKeyboard(LPKMX_KEYBOARD kbd, KMX_CHAR* filename);
 
 PKMX_WCHAR KMX_incxstr(PKMX_WCHAR p);
+
+// _S2 Open files on windows and non-windows platforms. Datatypes for Filename and mode must be the same.
+// return FILE* if file could be opened; FILE must to be closed in calling function
+FILE* Open_File(const KMX_CHAR* Filename, const KMX_CHAR* mode);
+FILE* Open_File(const KMX_WCHAR* Filename, const KMX_WCHAR* mode);
 
 #endif // _KMXFILE_H
